@@ -7,21 +7,22 @@
     "start": "ts-node src/main"
   },
   "dependencies": {
-    "@nestjs/common": "6.1.1",
-    "@nestjs/core": "6.1.1",
-    "@nestjs/microservices": "6.1.1",
-    "@nestjs/testing": "6.1.1",
-    "@nestjs/websockets": "6.1.1",
-    "amqp-connection-manager": "2.3.1",
-    "class-transformer": "0.2.0",
+    "@nestjs/common": "6.3.1",
+    "@nestjs/core": "6.3.1",
+    "@nestjs/microservices": "6.3.1",
+    "@nestjs/platform-express": "6.3.1",
+    "@nestjs/testing": "6.3.1",
+    "@nestjs/websockets": "6.3.1",
+    "amqp-connection-manager": "2.3.2",
+    "class-transformer": "0.2.3",
     "class-validator": "0.9.1",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.4.0",
-    "typescript": "3.4.4"
+    "rxjs": "6.5.2",
+    "typescript": "3.5.1"
   },
   "devDependencies": {
-    "@types/node": "7.10.5",
+    "@types/node": "7.10.6",
     "supertest": "4.0.2",
-    "ts-node": "8.1.0"
+    "ts-node": "8.2.0"
   }
 }
