@@ -9,18 +9,18 @@
     "start:prod": "node dist/main.js"
   },
   "dependencies": {
-    "@nestjs/common": "6.1.1",
-    "@nestjs/core": "6.1.1",
+    "@nestjs/common": "6.3.1",
+    "@nestjs/core": "6.3.1",
     "@nestjs/mongoose": "6.1.2",
-    "mongoose": "5.5.2",
+    "mongoose": "5.5.13",
     "reflect-metadata": "0.1.13",
-    "rxjs": "6.4.0",
-    "typescript": "3.4.4"
+    "rxjs": "6.5.2",
+    "typescript": "3.5.1"
   },
   "devDependencies": {
-    "@types/mongoose": "5.3.26",
-    "@types/node": "7.10.5",
-    "ts-node": "8.1.0",
-    "tslint": "5.16.0"
+    "@types/mongoose": "5.5.5",
+    "@types/node": "7.10.6",
+    "ts-node": "8.2.0",
+    "tslint": "5.17.0"
   }
 }
