@@ -1,5 +1,12 @@
 import { Body, Controller, HttpCode, Post } from '@nestjs/common';
-import { Client, ClientGrpc, GrpcMethod, Transport } from '@nestjs/microservices';
+import {
+  Client,
+  ClientGrpc,
+  GrpcMethod,
+  GrpcStreamCall,
+  GrpcStreamMethod,
+  Transport,
+} from '@nestjs/microservices';
 import { join } from 'path';
 import { Observable, of } from 'rxjs';
 
@@ -25,6 +32,29 @@ export class GrpcController {
   async sum({ data }: { data: number[] }): Promise<any> {
     return of({
       result: data.reduce((a, b) => a + b),
+    });
+  }
+
+  @GrpcStreamMethod('Math')
+  async sumStream(messages: Observable<any>): Promise<any> {
+    return new Promise<any>((resolve, reject) => {
+      messages.subscribe(
+        msg => {
+          resolve({
+            result: msg.data.reduce((a, b) => a + b),
+          });
+        },
+        err => {
+          reject(err);
+        },
+      );
+    });
+  }
+
+  @GrpcStreamCall('Math')
+  async sumStreamPass(stream: any) {
+    stream.on('data', (msg: any) => {
+      stream.write({ result: msg.data.reduce((a, b) => a + b) });
     });
   }
 }
