@@ -6,7 +6,8 @@ export type Provider<T = any> =
   | Type<any>
   | ClassProvider<T>
   | ValueProvider<T>
-  | FactoryProvider<T>;
+  | FactoryProvider<T>
+  | ExistingProvider<T>;
 
 export interface ClassProvider<T = any> {
   provide: string | symbol | Type<any> | Abstract<any> | Function;
@@ -22,6 +23,11 @@ export interface ValueProvider<T = any> {
 export interface FactoryProvider<T = any> {
   provide: string | symbol | Type<any> | Abstract<any> | Function;
   useFactory: (...args: any[]) => T;
-  inject?: Array<Type<any> | string | any>;
+  inject?: Array<Type<any> | string | symbol | Abstract<any> | Function>;
   scope?: Scope;
+}
+
+export interface ExistingProvider<T = any> {
+  provide: string | symbol | Type<any> | Abstract<any> | Function;
+  useExisting: any;
 }
