@@ -1,2 +1,3 @@
-export * from './validation.pipe';
 export * from './parse-int.pipe';
+export * from './parse-uuid.pipe';
+export * from './validation.pipe';
