@@ -1,6 +1,10 @@
 import { Optional } from '../decorators';
 import { Injectable } from '../decorators/core';
-import { ArgumentMetadata, BadRequestException, ValidationError } from '../index';
+import {
+  ArgumentMetadata,
+  BadRequestException,
+  ValidationError,
+} from '../index';
 import { ClassTransformOptions } from '../interfaces/external/class-transform-options.interface';
 import { ValidatorOptions } from '../interfaces/external/validator-options.interface';
 import { PipeTransform } from '../interfaces/features/pipe-transform.interface';
@@ -12,6 +16,7 @@ export interface ValidationPipeOptions extends ValidatorOptions {
   disableErrorMessages?: boolean;
   transformOptions?: ClassTransformOptions;
   exceptionFactory?: (errors: ValidationError[]) => any;
+  validateCustomDecorators?: boolean;
 }
 
 let classValidator: any = {};
@@ -24,6 +29,7 @@ export class ValidationPipe implements PipeTransform<any> {
   protected validatorOptions: ValidatorOptions;
   protected transformOptions: ClassTransformOptions;
   protected exceptionFactory: (errors: ValidationError[]) => any;
+  protected validateCustomDecorators: boolean;
 
   constructor(@Optional() options?: ValidationPipeOptions) {
     options = options || {};
@@ -31,12 +37,14 @@ export class ValidationPipe implements PipeTransform<any> {
       transform,
       disableErrorMessages,
       transformOptions,
+      validateCustomDecorators,
       ...validatorOptions
     } = options;
     this.isTransformEnabled = !!transform;
     this.validatorOptions = validatorOptions;
     this.transformOptions = transformOptions;
     this.isDetailedOutputDisabled = disableErrorMessages;
+    this.validateCustomDecorators = validateCustomDecorators || false;
     this.exceptionFactory =
       options.exceptionFactory ||
       (errors =>
@@ -57,10 +65,13 @@ export class ValidationPipe implements PipeTransform<any> {
     if (!metatype || !this.toValidate(metadata)) {
       return value;
     }
+    value = this.toEmptyIfNil(value);
+
+    this.stripProtoKeys(value);
     const entity = classTransformer.plainToClass(
       metatype,
-      this.toEmptyIfNil(value),
-      this.transformOptions
+      value,
+      this.transformOptions,
     );
     const errors = await classValidator.validate(entity, this.validatorOptions);
     if (errors.length > 0) {
@@ -69,20 +80,28 @@ export class ValidationPipe implements PipeTransform<any> {
     return this.isTransformEnabled
       ? entity
       : Object.keys(this.validatorOptions).length > 0
-        ? classTransformer.classToPlain(entity, this.transformOptions)
-        : value;
+      ? classTransformer.classToPlain(entity, this.transformOptions)
+      : value;
   }
 
   private toValidate(metadata: ArgumentMetadata): boolean {
     const { metatype, type } = metadata;
-    if (type === 'custom') {
+    if (type === 'custom' && !this.validateCustomDecorators) {
       return false;
     }
     const types = [String, Boolean, Number, Array, Object];
     return !types.some(t => metatype === t) && !isNil(metatype);
   }
 
-  toEmptyIfNil<T = any, R = any>(value: T): R | {} {
+  private toEmptyIfNil<T = any, R = any>(value: T): R | {} {
     return isNil(value) ? {} : value;
+  }
+
+  private stripProtoKeys(value: Record<string, any>) {
+    delete value.__proto__;
+    const keys = Object.keys(value);
+    keys
+      .filter(key => typeof value[key] === 'object' && value[key])
+      .forEach(key => this.stripProtoKeys(value[key]));
   }
 }
