@@ -16,6 +16,7 @@ import {
   ClassProvider,
   FactoryProvider,
   ValueProvider,
+  ExistingProvider,
 } from '@nestjs/common/interfaces';
 import { Controller } from '@nestjs/common/interfaces/controllers/controller.interface';
 import { Injectable } from '@nestjs/common/interfaces/injectable.interface';
@@ -187,14 +188,15 @@ export class DependenciesScanner {
       component.prototype,
       this.reflectKeyMetadata.bind(this, component, metadataKey),
     );
-    const flattenMethodsInjectables = methodsInjectables.reduce<any[]>(
-      (a: any[], b) => a.concat(b),
+    const flattenMethodsInjectables = methodsInjectables.reduce(
+      (a: any[], b: any[]) => a.concat(b),
       [],
-    );
-    const injectables = [
+    ) as any[];
+    const combinedInjectables = [
       ...controllerInjectables,
       ...flattenMethodsInjectables,
     ].filter(isFunction);
+    const injectables = Array.from(new Set(combinedInjectables));
 
     injectables.forEach(injectable =>
       this.insertInjectable(injectable, token, component),
@@ -255,7 +257,11 @@ export class DependenciesScanner {
 
   public isCustomProvider(
     provider: Provider,
-  ): provider is ClassProvider | ValueProvider | FactoryProvider {
+  ): provider is
+    | ClassProvider
+    | ValueProvider
+    | FactoryProvider
+    | ExistingProvider {
     return provider && !isNil((provider as any).provide);
   }
 
@@ -266,8 +272,11 @@ export class DependenciesScanner {
     }
     const applyProvidersMap = this.getApplyProvidersMap();
     const providersKeys = Object.keys(applyProvidersMap);
-    const type = (provider as ClassProvider | ValueProvider | FactoryProvider)
-      .provide;
+    const type = (provider as
+      | ClassProvider
+      | ValueProvider
+      | FactoryProvider
+      | ExistingProvider).provide;
 
     if (!providersKeys.includes(type as string)) {
       return this.container.addProvider(provider as any, token);
