@@ -1,16 +1,19 @@
-import * as sinon from 'sinon';
 import { expect } from 'chai';
+import * as sinon from 'sinon';
 import { ExceptionsZone } from '../../../errors/exceptions-zone';
-import { UNHANDLED_RUNTIME_EXCEPTION } from '../../../errors/messages';
 
 describe('ExceptionsZone', () => {
+  const rethrow = err => {
+    throw err;
+  };
+
   describe('run', () => {
     let callback: sinon.SinonSpy;
     beforeEach(() => {
       callback = sinon.spy();
     });
     it('should call callback', () => {
-      ExceptionsZone.run(callback as any);
+      ExceptionsZone.run(callback as any, rethrow);
       expect(callback.called).to.be.true;
     });
     describe('when callback throws exception', () => {
@@ -18,17 +21,15 @@ describe('ExceptionsZone', () => {
         handle: () => {},
       };
       let handleSpy: sinon.SinonSpy;
-      beforeEach(() => {
+      before(() => {
         (ExceptionsZone as any).exceptionHandler = exceptionHandler;
         handleSpy = sinon.spy(exceptionHandler, 'handle');
       });
-      it('should call "handle" method of exceptionHandler and throws UNHANDLED_RUNTIME_EXCEPTION', () => {
+      it('should call "handle" method of exceptionHandler and rethrows', () => {
         const throwsCallback = () => {
-          throw 3;
+          throw new Error('');
         };
-        expect(() => ExceptionsZone.run(throwsCallback)).to.throws(
-          UNHANDLED_RUNTIME_EXCEPTION,
-        );
+        expect(() => ExceptionsZone.run(throwsCallback, rethrow)).to.throws();
         expect(handleSpy.called).to.be.true;
       });
     });
@@ -39,7 +40,7 @@ describe('ExceptionsZone', () => {
       callback = sinon.spy();
     });
     it('should call callback', async () => {
-      await ExceptionsZone.asyncRun(callback as any);
+      await ExceptionsZone.asyncRun(callback as any, rethrow);
       expect(callback.called).to.be.true;
     });
     describe('when callback throws exception', () => {
@@ -47,16 +48,16 @@ describe('ExceptionsZone', () => {
         handle: () => {},
       };
       let handleSpy: sinon.SinonSpy;
-      beforeEach(() => {
+      before(() => {
         (ExceptionsZone as any).exceptionHandler = exceptionHandler;
         handleSpy = sinon.spy(exceptionHandler, 'handle');
       });
-      it('should call "handle" method of exceptionHandler and throws UNHANDLED_RUNTIME_EXCEPTION', async () => {
+      it('should call "handle" method of exceptionHandler and rethrows error', async () => {
         const throwsCallback = () => {
-          throw 3;
+          throw new Error('');
         };
-        expect(ExceptionsZone.asyncRun(throwsCallback)).to.eventually.be
-          .rejected;
+        expect(ExceptionsZone.asyncRun(throwsCallback, rethrow)).to.eventually
+          .be.rejected;
       });
     });
   });
