@@ -2,12 +2,16 @@ import { Logger } from '@nestjs/common/services/logger.service';
 import { loadPackage } from '@nestjs/common/utils/load-package.util';
 import { isObject } from '@nestjs/common/utils/shared.utils';
 import { Observable } from 'rxjs';
-import { GRPC_DEFAULT_PROTO_LOADER, GRPC_DEFAULT_URL } from '../constants';
+import {
+  GRPC_DEFAULT_MAX_RECEIVE_MESSAGE_LENGTH,
+  GRPC_DEFAULT_MAX_SEND_MESSAGE_LENGTH,
+  GRPC_DEFAULT_PROTO_LOADER,
+  GRPC_DEFAULT_URL,
+} from '../constants';
 import { InvalidGrpcPackageException } from '../errors/invalid-grpc-package.exception';
 import { InvalidGrpcServiceException } from '../errors/invalid-grpc-service.exception';
 import { InvalidProtoDefinitionException } from '../errors/invalid-proto-definition.exception';
 import { ClientGrpc, GrpcOptions } from '../interfaces';
-import { ClientOptions } from '../interfaces/client-metadata.interface';
 import { ClientProxy } from './client-proxy';
 import { GRPC_CANCELLED } from './constants';
 
@@ -19,14 +23,12 @@ export class ClientGrpcProxy extends ClientProxy implements ClientGrpc {
   protected readonly url: string;
   protected grpcClient: any;
 
-  constructor(protected readonly options: ClientOptions['options']) {
+  constructor(protected readonly options: GrpcOptions['options']) {
     super();
-    this.url =
-      this.getOptionsProp<GrpcOptions>(options, 'url') || GRPC_DEFAULT_URL;
+    this.url = this.getOptionsProp(options, 'url') || GRPC_DEFAULT_URL;
 
     const protoLoader =
-      this.getOptionsProp<GrpcOptions>(options, 'protoLoader') ||
-      GRPC_DEFAULT_PROTO_LOADER;
+      this.getOptionsProp(options, 'protoLoader') || GRPC_DEFAULT_PROTO_LOADER;
 
     grpcPackage = loadPackage('grpc', ClientGrpcProxy.name, () =>
       require('grpc'),
@@ -36,9 +38,29 @@ export class ClientGrpcProxy extends ClientProxy implements ClientGrpc {
   }
 
   public getService<T extends {}>(name: string): T {
+    const maxSendMessageLengthKey = 'grpc.max_send_message_length';
+    const maxReceiveMessageLengthKey = 'grpc.max_receive_message_length';
+    const maxMessageLengthOptions = {
+      [maxSendMessageLengthKey]: this.getOptionsProp(
+        this.options,
+        'maxSendMessageLength',
+        GRPC_DEFAULT_MAX_SEND_MESSAGE_LENGTH,
+      ),
+      [maxReceiveMessageLengthKey]: this.getOptionsProp(
+        this.options,
+        'maxReceiveMessageLength',
+        GRPC_DEFAULT_MAX_RECEIVE_MESSAGE_LENGTH,
+      ),
+    };
     const options: any = isObject(this.options)
-      ? { ...this.options, loader: '' }
-      : {};
+      ? {
+          ...this.options,
+          ...maxMessageLengthOptions,
+          loader: '',
+        }
+      : {
+          ...maxMessageLengthOptions,
+        };
 
     if (!this.grpcClient[name]) {
       throw new InvalidGrpcServiceException();
@@ -119,10 +141,7 @@ export class ClientGrpcProxy extends ClientProxy implements ClientGrpc {
 
   public createClient(): any {
     const grpcContext = this.loadProto();
-    const packageName = this.getOptionsProp<GrpcOptions>(
-      this.options,
-      'package',
-    );
+    const packageName = this.getOptionsProp(this.options, 'package');
     const grpcPkg = this.lookupPackage(grpcContext, packageName);
     if (!grpcPkg) {
       const invalidPackageError = new InvalidGrpcPackageException();
@@ -134,8 +153,8 @@ export class ClientGrpcProxy extends ClientProxy implements ClientGrpc {
 
   public loadProto(): any {
     try {
-      const file = this.getOptionsProp<GrpcOptions>(this.options, 'protoPath');
-      const loader = this.getOptionsProp<GrpcOptions>(this.options, 'loader');
+      const file = this.getOptionsProp(this.options, 'protoPath');
+      const loader = this.getOptionsProp(this.options, 'loader');
 
       const packageDefinition = grpcProtoLoaderPackage.loadSync(file, loader);
       const packageObject = grpcPackage.loadPackageDefinition(
